requirements.txt
PyYAML==5.1
pytest==4.5.0
pytest-cov==2.7.1
python-interface==1.5.1
statistics==1.0.3.5
parameterized==0.7.0